+++ packages-rc/rc-keyboard-mac/src/index.ts
@@ -2,10 +2,10 @@ export { default } from './with-model';
 
 export {
   KeyboardCode
-} from '@kcuf/rc-headless-keyboard-mac';
+} from '@kcuf/rc-keyboard-mac-headless';
 
 export type {
   KeyData,
   KeyboardModifiers,
   KeyboardProps
-} from '@kcuf/rc-headless-keyboard-mac';
+} from '@kcuf/rc-keyboard-mac-headless';