+++ packages-rc/rc-keyboard-mac/src/ui/rc/keyboard-key/index.tsx
@@ -9,7 +9,7 @@ import {
   useIsKeyActive,
   useIsKeyOn,
   useHandleKeyClick
-} from '@kcuf/rc-headless-keyboard-mac';
+} from '@kcuf/rc-keyboard-mac-headless';
 
 import {
   KEY_HEIGHT,