+++ packages-rc/rc-keyboard-mac/src/ui/util/get-keyboard-key-display-name.ts
@@ -1,6 +1,6 @@
 import {
   KeyData
-} from '@kcuf/rc-headless-keyboard-mac';
+} from '@kcuf/rc-keyboard-mac-headless';
 
 export default function getKeyboardKeyDisplayName(data: KeyData): string | [string, string] {
   if (data.name) {